requirements.txt
torch
torchvision
medmnist
matplotlib
pillow